Server/requirements.txt
faker
requests
oauthlib
cryptography
django
djangorestframework
django-cors-headers
django-environ==0.12.0
django-auditlog==3.4.1
django-allauth==65.13.1
djangorestframework-simplejwt
psycopg[c]==3.3.2
argparse
